Stack Operation
A slack is a si mple data struc ture used for s toring dnla (similar to Linked Lists). In a stack, the o rder in which
the data arrives is impona nt. A pile of pln tcs in a cafeteria is a good example of n stuck . The plates arc added to
the stack as they arc cleaned and they a rc placed on the top. When a plate, is required it is taken from the top of
the stack. The first pla te placed on the stack is the last one lo be used.
Definition: A stack is an ordered list in which insertion and deletion a re done al one end, ca lled top. The last
c le ment inse rted is the first one to be deleted. Hence, il is called the Last in First out (LI PO) or F'irst in Last oul
(f'I LO) list.